chore: order tracking and payment successful

+++ api/controller/cart.controller.js
@@ -4,32 +4,54 @@ router.post("/", async (req, res) => {
   try {
     const { user, cartItems } = req.body;
 
-    // Create a new cart for the user
-    const { id: cartId } = await cartService.createCart(user);
+    // Check if the user already has a cart
+    const existingCart = await cartService.getCartByUser(user);
 
-    // Modify cartItems to include the cartId
-    const modifiedCartItems = cartItems.items.map((item) => ({
-      ...item,
-      cartId: cartId,
-    }));
+    if (existingCart) {
+      // If the user has an existing cart, add cart items to it
+      const modifiedCartItems = cartItems.items.map((item) => ({
+        ...item,
+        cartId: existingCart.id,
+      }));
 
-    // Create cart items associated with the newly created cart
-    const createdCartItems = await cartItemService.createCartItem(
-      modifiedCartItems
-    );
+      const createdCartItems = await cartItemService.createCartItem(
+        modifiedCartItems
+      );
 
-    res.json({
-      code: 200,
-      message: "Cart and cart items created successfully",
-      meta: createdCartItems,
-    });
+      res.json({
+        code: 200,
+        message: "Cart items added to the existing cart",
+        meta: createdCartItems,
+      });
+    } else {
+      // If the user does not have an existing cart, create a new cart and add cart items
+      const { id: cartId } = await cartService.createCart(user);
+
+      const modifiedCartItems = cartItems.items.map((item) => ({
+        ...item,
+        cartId: cartId,
+      }));
+
+      const createdCartItems = await cartItemService.createCartItem(
+        modifiedCartItems
+      );
+
+      res.json({
+        code: 200,
+        message: "Cart and cart items created successfully",
+        meta: createdCartItems,
+      });
+    }
   } catch (error) {
     console.error(error);
-    res
-      .status(500)
-      .json({ code: 500, message: "Internal Server Error", meta: null });
+    res.status(500).json({
+      code: 500,
+      message: "Internal Server Error",
+      meta: null,
+    });
   }
 });
+
 router.put("/", async (req, res) => {
   try {
     const { id, quantity } = req.body;
@@ -55,17 +77,55 @@ router.put("/", async (req, res) => {
       .json({ code: 500, message: "Internal Server Error", meta: null });
   }
 });
+
 router.delete("/:id", async (req, res) => {
   try {
     const { id } = req.params;
+    const { cartItemIds } = req.body;
 
-    // Call the deleteCart method from the service
-    const deletedCart = await cartService.deleteCart(id);
+    // Delete each cart item using the CartItemService
+    for (const cartItemId of cartItemIds) {
+      await cartItemService.deleteCartItem(cartItemId);
+    }
+
+    // After deleting cart items, check if the cart is empty
+    const cart = await cartService.getCartById(id);
+
+    if (cart.cartItems.length === 0) {
+      // If there are no more cart items, delete the cart
+      await cartService.deleteCart(id);
+
+      res.json({
+        code: 200,
+        message: "Cart and cart items deleted successfully",
+        meta: null,
+      });
+    } else {
+      res.json({
+        code: 200,
+        message: "Cart items deleted successfully",
+        meta: null,
+      });
+    }
+  } catch (error) {
+    console.error(error);
+    res
+      .status(500)
+      .json({ code: 500, message: "Internal Server Error", meta: null });
+  }
+});
+
+router.get("/:id/cartitems", async (req, res) => {
+  try {
+    const { id } = req.params;
+
+    // Call the getCartItemsByCartId method from the CartService
+    const cartItems = await cartService.getCartItemsByCartId(id);
 
     res.json({
       code: 200,
-      message: "Cart deleted successfully",
-      meta: deletedCart,
+      message: "Cart items fetched successfully",
+      meta: cartItems,
     });
   } catch (error) {
     console.error(error);
